api/pyproject.toml
[tool.poetry]
name = "heygpt"
version = "0.1.0"
description = ""
authors = ["Rishang <rishang@localhost.com>"]
readme = "README.md"
packages = [
  { include = "heygpt" }
]

[tool.poetry.scripts]
heygpt = "heygpt.cli:app"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
fastapi = {extras = ["all"], version = "^0.95.2"}
uvicorn = "^0.22.0"
typer = "^0.9.0"
python-dotenv = "^1.0.0"
bardapi = "^0.1.11"
openai = "^0.27.8"
rich = "^13.4.1"
prompt-toolkit = "^3.0.38"
streamlit = "^1.23.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.2"
black = "^23.3.0"
mypy = "^1.3.0"
types-requests = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
